requirements.txt
streamlit>=1.30.0
bcrypt>=4.1.2
PyPDF2>=3.0.1
python-docx>=1.1.0
pandas>=2.1.4
plotly>=5.18.0
fpdf2>=2.7.7
Pillow>=10.2.0
google-generativeai>=0.3.0
requests>=2.31.0
